chore(engine): move config.py from core/ to top-level
