+++ packages/devqubit-engine/tests/conftest.py
@@ -12,7 +12,7 @@ from typing import Any, Callable
 
 import pytest
 from devqubit_engine.circuit.models import SDK, CircuitData, CircuitFormat
-from devqubit_engine.core.config import Config
+from devqubit_engine.config import Config
 from devqubit_engine.core.record import RunRecord
 from devqubit_engine.storage.backends.local import LocalRegistry, LocalStore
 from devqubit_engine.storage.types import ArtifactRef
